회원 레파지토리 개발
