+++ src/main/java/com/hackathonOne/hackathon/domain/entity/Allergy.java
@@ -1,4 +1,4 @@
-package com.hackathonOne.hackathon.domain;
+package com.hackathonOne.hackathon.domain.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
@@ -6,11 +6,10 @@ import lombok.Setter;
 
 @Entity
 @Getter @Setter
-public class SpecialType {
+public class Allergy {
 
-    @Id
-    @GeneratedValue
-    @Column(name="specialType_id")
+    @Id @GeneratedValue
+    @Column(name="allergy_id")
     private Long id;
 
     private String name;
@@ -18,4 +17,5 @@ public class SpecialType {
     @ManyToOne
     @JoinColumn(name="member_id")
     private Member member;
+
 }
