+++ src/main/java/com/hackathonOne/hackathon/domain/entity/CanEat.java
@@ -1,4 +1,4 @@
-package com.hackathonOne.hackathon.domain;
+package com.hackathonOne.hackathon.domain.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
