+++ src/main/java/com/hackathonOne/hackathon/repository/MemberRepository.java
@@ -1,0 +1,20 @@
+package com.hackathonOne.hackathon.repository;
+
+import com.hackathonOne.hackathon.domain.entity.Member;
+import jakarta.persistence.EntityManager;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+
+@Repository
+@RequiredArgsConstructor
+public class MemberRepository {
+    private final EntityManager em;
+
+    public void save(Member member){
+        em.persist(member);
+    }
+
+    public Member findOne(Long id) { return em.find(Member.class , id); }
+
+
+}
